requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.1
certifi==2018.1.18
chardet==3.0.4
Click==7.0
configparser==3.5.0
cryptography==2.1.4
docker==2.5.1
docker-compose==1.17.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.7.17
enum34==1.1.6
fasteners==0.12.0
Flask==1.0.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
idna==2.6
ipaddress==1.0.17
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.3.1
lockfile==0.12.2
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
monotonic==1.0
mysql-connector-python==8.0.13
mysqlclient==1.3.10
pbr==3.1.1
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pygobject==3.26.1
pylint==1.9.3
pyOpenSSL==17.5.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.13
texttable==0.9.1
urllib3==1.22
virtualenv==16.1.0
websocket-client==0.44.0
Werkzeug==0.14.1
wrapt==1.10.11
